requirements.txt
opencv-python
dlib
Pillow
numpy
imutils
argparse